Crate pallet_crowdloan_reward

Crowdloan reward pallet

This pallet implements a specific rewarding strategy for crowdloan campaign. It worth pointing out that this pallet works hand in hand with the [pallet-crowdloan-claim] pallet, the latter being responsible for managing reward claims (including security aspects, such as malicious Denial of Service or replay attacks).

[Config]

[Call]

[Pallet]

[Reward](pallet_crowdloan_claim::traits::Reward)

Overview

The function of this pallet is to provide the Centrifuge-specific reward functionality for contributors of the relay chain crowdloan. In order to provide this functionality the pallet implements the Reward Trait from the Claim Pallet. Before any rewards can be provided to contributors the pallet MUST be initialized, so that the modules account holds the necessary funds to reward. All rewards are payed in form of a vested_transfer with a fixed vesting time. The vesting time is defined via the modules Config trait.

Terminology

For information on terms and concepts used in this pallet, please refer to the pallet’ specification document.

Goals

Usage

Interface

Supported Origins

Valid origin is an administrator or root.

Types

Events

Errors

Dispatchable Functions

[set_vesting_start] : Origin must be admin or root, and allows to set the start of the vesting after the initialization.

[set_vesting_period] : Origin must be admin or root, and allows to set the period of the vesting after the initialization.

[set_conversion_rate] : Origin must be admin or root, and allows to set the conversion rate between relay chain and native balance after the initialization.

[set_direct_payout_ratio] : Origin must be admin or root, and allows to set the ratio between vested and direct payout amount after the initialization.

Public Functions

Genesis Configuration

Dependencies

This pallet works hand in hand with [pallet-crowdloan-claim] pallet. In fact, it must implement this pallet’s [pallet-crowdloan-claim::traits::Reward] trait so that to interact.

References

Credits

Frederik Schulz frederik@centrifuge.io

Re-exports

pub use pallet::*;

pub use crate::weights::WeightInfo;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_crowdloan_reward