Crate pallet_crowdloan_reward

Crowdloan reward pallet

This pallet implements a specific rewarding strategy for crowdloan campaign. It worth pointing out that this pallet works hand in hand with the [pallet-crowdloan-claim] pallet, the latter being responsible for managing reward claims (including security aspects, such as malicious Denial of Service or replay attacks).

Overview

The function of this pallet is to provide the Centrifuge-specific reward functionality for contributors of the relay chain crowdloan. In order to provide this functionality the pallet implements the Reward Trait from the Claim Pallet. Before any rewards can be provided to contributors the pallet MUST be initialized, so that the modules account holds the necessary funds to reward. All rewards are payed in form of a vested_transfer with a fixed vesting time. The vesting time is defined via the modules Config trait.

Terminology

For information on terms and concepts used in this pallet, please refer to the pallet’ specification document.

Dependencies

This pallet works hand in hand with [pallet-crowdloan-claim] pallet. In fact, it must implement this pallet’s [pallet-crowdloan-claim::traits::Reward] trait so that to interact.

References

Credits

Frederik Schulz frederik@centrifuge.io

Re-exports

pub use crate::weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights